+++ src/board.rs
@@ -348,7 +348,7 @@ macro_rules! board_split_flash {
 
 #[macro_export]
 macro_rules! board_split_sdram {
-    ($size:expr, $cp:expr, $dp:expr, $ccdr:expr, $pins:expr) => {{
+    ($cp:expr, $dp:expr, $ccdr:expr, $pins:expr) => {{
         use daisy::hal::delay::DelayFromCountDownTimer;
         use daisy::hal::prelude::*;
         let mut delay = DelayFromCountDownTimer::new($dp.TIM3.timer(
@@ -357,7 +357,6 @@ macro_rules! board_split_sdram {
             &$ccdr.clocks,
         ));
         let sdram = daisy::sdram::SDRAM::new(
-            $size,
             $pins.SDRAM,
             &$ccdr.clocks,
             $dp.FMC,